refactor move course router -> course folder
