+++ learnhub_backend/main.py
@@ -4,9 +4,11 @@ import uvicorn
 
 from .database import db_client
 from .program.router import router as program_router
+from .program.course.router import router as course_router
 
 app = FastAPI()
 app.include_router(program_router)
+app.include_router(course_router)
 
 
 try:
