+++ learnhub_backend/program/config.py
@@ -1,1 +1,1 @@
-
+# ROUTE SPECIFIC CONFIG
